+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2021-10-19
+### Added
+- Typescript types support by [@aghontpi](https://github.com/aghontpi).
+- JSDoc comments for code.
+- Quote object now has success property to denote status.
+- `getRandomQuoteByAnime` and `getRandomQuoteByCharacter` functions.
+- Substring search.
+
+### Fixed
+- `randomQuote` out of bounds error.
+
 ## [1.1.2] - 2020-04-07
 ### Security
 - [Bumped standard from 14.3.1 to 14.3.3 to fix prototype pollution in minimist](https://github.com/advisories/GHSA-vh95-rmgr-6w4m).
@@ -30,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release with 8510 quotes.
 
 [unreleased]: https://github.com/FR0ST1N/animequotes/compare/master...develop
+[2.0.0]: https://github.com/FR0ST1N/animequotes/compare/v1.1.2...v2.0.0
 [1.1.2]: https://github.com/FR0ST1N/animequotes/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/FR0ST1N/animequotes/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/FR0ST1N/animequotes/compare/v1.0.0...v1.1.0